Separate test into test and test_subprocess
