+++ rules/scala/test_common/AnxLogger.scala
@@ -1,0 +1,22 @@
+package annex
+
+import sbt.testing.Logger
+
+final class AnxLogger(color: Boolean, verbosity: String) extends Logger with Serializable {
+  def ansiCodesSupported = color
+
+  def error(msg: String) = println(s"$msg")
+
+  def warn(msg: String) = println(s"$msg")
+
+  def info(msg: String) = verbosity match {
+    case "HIGH" | "MEDIUM" => println(s"$msg")
+    case _                 =>
+  }
+
+  def debug(msg: String) = verbosity match {
+    case "HIGH" => println(s"$msg")
+  }
+
+  def trace(err: Throwable) = println(s"${err.getMessage}")
+}
