+++ rules/scala/test_common/Fingerprint.scala
