+++ rules/scala/test_common/TestRequest.scala
@@ -1,0 +1,11 @@
+package annex
+
+import sbt.testing.Logger
+
+class TestRequest(
+  val framework: String,
+  val test: TestDefinition,
+  val scopeAndTestName: String,
+  val classpath: Seq[String],
+  val logger: Logger with Serializable
+) extends Serializable
